+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@
 * Added `slices.CloseToEpsilon` to easily customize tests.
 * Fixed `Variable.SetValuePreservingOld` when shapes change.
 * `Scatter` doesn't assume indices are sorted or unique.
+* Plotly training plots: added `WithCustomMetricFn` for custom metrics and `ScheduleEveryNSteps`.
 
 ## 0.9.1 - 2024/04/19